Trying to do things(TM)
will have to make sure my logic is correct

+++ Neural Network/func.h
@@ -5,7 +5,7 @@ namespace nnet {
 
     class IFunc {
     public:
-        virtual float apply(float net) = 0;
+        virtual double apply(double net) = 0;
     };
 }
