+++ Neural Network/functions.cpp
@@ -2,15 +2,15 @@
 #include <math.h>
 
 namespace nnet {
-    float Sigmoid::apply(float net) {
+    double Sigmoid::apply(double net) {
         return 1/(1 + exp(-net*(1))); //1 is lambda I may need to add functionality to control it
     }
     
-    float Linear::apply(float net) {
+    double Linear::apply(double net) {
         return net;
     }
 
-    float Ramp::apply(float net) {
+    double Ramp::apply(double net) {
         if (net < -1)
             return 0;
         else if (net > 1)
@@ -19,7 +19,7 @@ namespace nnet {
             return net;
     }
     
-    float Step::apply(float net) {
+    double Step::apply(double net) {
         return net < 0 ? 0 : 1;
     }
 }
