+++ Neural Network/neurIO.h
@@ -15,10 +15,10 @@ using std::stof;
 namespace nnet
 {
     vector<string> getData(const string &filename);
-    vector<vector<float> > getDataCsv(const string &filename);
-    void printData(const vector<vector<float> > &data);
+    vector<vector<double> > getDataCsv(const string &filename);
+    void printData(const vector<vector<double> > &data);
     vector<string> &split(const string &s, char delim, vector<string> &elems);
     vector<string> split(const string &s, char delim);    
-    vector<vector<float> > getDataCsv(const ifstream &stream);
+    vector<vector<double> > getDataCsv(const ifstream &stream);
 }
 #endif /* NEURIO_H */
